requirements_dev.txt
boto3==1.4.4
botocore==1.5.52
docutils==0.13.1
jmespath==0.9.2
psycopg2==2.7.3.2
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
s3transfer==0.1.10
SQLAlchemy==1.1.9
sqlalchemy-redshift==0.6.0
numpy==1.14.0
pandas==0.22.0
pyarrow==0.8.0
